+++ src/components/EditMessage/EditForm.jsx
@@ -4,7 +4,7 @@ import styled from 'styled-components';
 import EditorBox from '../WritingMessage/TextEditor';
 import Dropdown from '../common/Dropdown';
 import { getMessage, getProfile } from '../../api/GetApi';
-import PutPatchButton from '../common/Buttons/PutPatchButton';
+import PutPatchButton from '../button/PutPatchButton';
 import { patchMessage, putMessage } from '../../api/PutPatchApi';
 import { bold24, regular16 } from '../../styles/fontStyle';
 import { DISPLAY_SIZE } from '../../constants/SIZE_SET';
