+++ src/components/button/DeleteMessageButton.jsx
@@ -1,19 +1,19 @@
 import styled from 'styled-components';
 import { useLocation } from 'react-router-dom';
-import { Outlined36 } from '../../../styles/ButtonStyle';
+import { Outlined36 } from '../../styles/ButtonStyle';
 
-function DeleteMessageButton({ id, onDelete }) {
+function DeleteMessageButton({ id: messageId, onDelete }) {
   const location = useLocation();
   const isEditRoute = location.pathname.includes('/edit');
 
   const handleButtonClick = async (e) => {
     e.stopPropagation();
-    await onDelete(id);
+    await onDelete(messageId);
   };
 
   return (
     <Button onClick={handleButtonClick} $isDisplay={isEditRoute}>
-      <DeleteImg src="/img/deleted.svg" alt="삭제" />
+      <DeleteImg src="/img/deleted.svg" alt="메세지 삭제" />
     </Button>
   );
 }
