+++ src/components/button/EditButton.jsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 import { useNavigate, useParams } from 'react-router-dom';
-import { Primary56 } from '../../../styles/ButtonStyle';
+import { Primary56 } from '../../styles/ButtonStyle';
 
 function EditButton() {
   const navigate = useNavigate();
