+++ src/pages/Post.jsx
@@ -2,7 +2,7 @@ import styled from 'styled-components';
 import { useEffect, useState } from 'react';
 import Header from '../components/common/Header';
 import SelectSection from '../components/post/postmake/SelectSection';
-import PostButton from '../components/common/Buttons/PostButton';
+import PostButton from '../components/button/PostButton';
 import PostInputSection from '../components/post/postmake/PostInputSection';
 import ColorSelection from '../components/post/postmake/ColorSelection';
 import ImageSelection from '../components/post/postmake/ImageSelection';
